fit_software/backend/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9